repackage, remove SpanDataModifier impl and rename builder
