+++ opentelemetry-android-instrumentation/src/test/java/io/opentelemetry/rum/internal/export/FilteringSpanExporterTest.java
@@ -1,10 +1,10 @@
-package io.opentelemetry.rum.export;
+package io.opentelemetry.rum.internal.export;
 
 import static org.junit.jupiter.api.Assertions.*;
 
 import org.junit.jupiter.api.Test;
 
-class AttributeModifyingSpanExporterTest {
+class FilteringSpanExporterTest {
 
     @Test
     void test() {
